tested on avx512
